refactor: move i18n lib to curve-ui-kit
