+++ apps/dao/src/components/PageVeCrv/components/FormLockDate.tsx
@@ -10,7 +10,7 @@ import { DEFAULT_FORM_EST_GAS } from '@/components/PageVeCrv/utils'
 import { REFRESH_INTERVAL } from '@/constants'
 import { getActiveStep, getStepStatus } from '@/ui/Stepper/helpers'
 import { formatDisplayDate, toCalendarDate } from '@/utils/utilsDates'
-import dayjs from '@/lib/dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 import networks from '@/networks'
 import usePageVisibleInterval from '@/hooks/usePageVisibleInterval'
 import useStore from '@/store/useStore'
