+++ apps/dao/src/pages/_app.tsx
@@ -12,7 +12,7 @@ import 'focus-visible'
 import '@/globals.css'
 import { ThemeProvider } from 'curve-ui-kit/src/shared/ui/ThemeProvider'
 
-import { dynamicActivate, initTranslation, updateAppLocale } from '@/lib/i18n'
+import { dynamicActivate, initTranslation, updateAppLocale } from '@ui-kit/lib/i18n'
 import { connectWalletLocales, initOnboard } from '@ui-kit/features/connect-wallet'
 import { getLocaleFromUrl, getStorageValue } from '@/utils'
 import { getIsMobile, getPageWidthClassName, isSuccess } from '@/ui/utils'
@@ -77,7 +77,10 @@ function CurveApp({ Component }: AppProps) {
     const { rLocale } = getLocaleFromUrl()
     const parsedLocale = rLocale?.value ?? 'en'
     initTranslation(i18n, parsedLocale)
-    dynamicActivate(parsedLocale)
+    ;(async () => {
+      let data = await import(`@/locales/${parsedLocale}/messages`)
+      dynamicActivate(parsedLocale, data)
+    })()
     updateAppLocale(parsedLocale, updateGlobalStoreByKey)
 
     // init onboard
