+++ apps/dao/src/store/createAppSlice.ts
@@ -1,7 +1,7 @@
 import type { GetState, SetState } from 'zustand'
 import type { State } from '@/store/useStore'
 import type { ConnectState } from '@/ui/utils'
-import type { Locale } from '@/lib/i18n'
+import type { Locale } from '@ui-kit/lib/i18n'
 
 import isEqual from 'lodash/isEqual'
 import produce from 'immer'
