+++ apps/lend/src/store/createAppSlice.ts
@@ -6,7 +6,7 @@ import { log } from '@ui-kit/lib/logging'
 import { setStorageValue } from '@/utils/utilsStorage'
 import isEqual from 'lodash/isEqual'
 import type { ThemeKey } from 'curve-ui-kit/src/themes/basic-theme'
-import { LocaleValue } from '@/lib/i18n'
+import { Locale } from '@ui-kit/lib/i18n'
 import { prefetchMarkets } from '@/entities/chain/chain-query'
 
 export type DefaultStateKeys = keyof typeof DEFAULT_STATE
@@ -23,7 +23,7 @@ type SliceState = {
   isLoadingCurve: true
   isMobile: boolean
   isPageVisible: boolean
-  locale: LocaleValue
+  locale: Locale['value']
   maxSlippage: string
   routerProps: RouterProps | null
   scrollY: number
