+++ apps/loan/src/hooks/usePageOnMount.ts
@@ -8,7 +8,7 @@ import { useCallback, useEffect } from 'react'
 import { getWalletSignerAddress, useConnectWallet, useSetChain, useSetLocale } from '@ui-kit/features/connect-wallet'
 
 import { CONNECT_STAGE, REFRESH_INTERVAL, ROUTE } from '@/constants'
-import { dynamicActivate, updateAppLocale } from '@/lib/i18n'
+import { dynamicActivate, updateAppLocale } from '@ui-kit/lib/i18n'
 import { getStorageValue, setStorageValue } from '@/utils/storage'
 import { getNetworkFromUrl, parseParams } from '@/utils/utilsRouter'
 import { getWalletChainId } from '@/store/createWalletSlice'
@@ -256,7 +256,10 @@ function usePageOnMount(params: Params, location: Location, navigate: NavigateFu
     if (isSuccess(connectState)) {
       const rLocale = parsedParams.rLocale?.value ?? 'en'
       if (rLocale !== document.documentElement.lang) {
-        dynamicActivate(rLocale)
+        ;(async () => {
+          let data = await import(`@/locales/${rLocale}/messages`)
+          dynamicActivate(rLocale, data)
+        })()
         updateAppLocale(rLocale, updateGlobalStoreByKey)
         updateWalletLocale(rLocale)
       } else if (
