+++ apps/loan/src/pages/_app.tsx
@@ -13,7 +13,7 @@ import { REFRESH_INTERVAL } from '@/constants'
 import GlobalStyle from '@/globalStyle'
 import usePageVisibleInterval from '@/hooks/usePageVisibleInterval'
 import Page from '@/layout/index'
-import { dynamicActivate, initTranslation } from '@/lib/i18n'
+import { dynamicActivate, initTranslation } from '@ui-kit/lib/i18n'
 import { messages as messagesEn } from '@/locales/en/messages.js'
 import networks from '@/networks'
 import { getPageWidthClassName } from '@/store/createLayoutSlice'
@@ -76,7 +76,10 @@ function CurveApp({ Component }: AppProps) {
     const { rLocale } = getLocaleFromUrl()
     const parsedLocale = rLocale?.value ?? 'en'
     initTranslation(i18n, parsedLocale)
-    dynamicActivate(parsedLocale)
+    ;(async () => {
+      let data = await import(`@/locales/${parsedLocale}/messages`)
+      dynamicActivate(parsedLocale, data)
+    })()
     updateGlobalStoreByKey('locale', parsedLocale)
 
     // init onboard
