+++ apps/loan/src/types/global.d.ts
@@ -1,7 +1,7 @@
 import { BrowserProvider } from 'ethers'
 import type { INetworkName } from '@curvefi/stablecoin-api/lib/interfaces'
 import type { LlammaTemplate } from '@curvefi/stablecoin-api/lib/llammas'
-import type { Locale } from '@/lib/i18n'
+import type { Locale } from '@ui-kit/lib/i18n'
 import type { NavigateFunction, Location, Params } from 'react-router'
 import type { ReactNode } from 'react'
 import type { TooltipProps } from '@/ui/Tooltip/types'
