+++ apps/loan/src/utils/storage.ts
@@ -1,5 +1,5 @@
 import merge from 'lodash/merge'
-import dayjs from 'dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 
 export const APP_STORAGE = {
   APP_CACHE: 'crvusd-app-cache',
