+++ apps/main/src/components/PageCrvLocker/components/FormLockCreate.tsx
@@ -11,7 +11,7 @@ import { formatDisplayDate, toCalendarDate } from '@/utils/utilsDates'
 import { formatNumber } from '@/ui/utils'
 import curvejsApi from '@/lib/curvejs'
 import usePageVisibleInterval from '@/hooks/usePageVisibleInterval'
-import dayjs from '@/lib/dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 import useStore from '@/store/useStore'
 import AlertFormError from '@/components/AlertFormError'
 import DetailInfoEstGas from '@/components/DetailInfoEstGas'
