+++ apps/main/src/components/PageCrvLocker/components/FormLockDate.tsx
@@ -9,7 +9,7 @@ import { REFRESH_INTERVAL } from '@/constants'
 import { getActiveStep, getStepStatus } from '@/ui/Stepper/helpers'
 import { formatDisplayDate, toCalendarDate } from '@/utils/utilsDates'
 import curvejsApi from '@/lib/curvejs'
-import dayjs from '@/lib/dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 import usePageVisibleInterval from '@/hooks/usePageVisibleInterval'
 import useStore from '@/store/useStore'
 import AlertBox from '@/ui/AlertBox'
