+++ apps/main/src/components/PagePool/PoolDetails/PoolParameters/index.tsx
@@ -6,7 +6,7 @@ import { breakpoints } from '@/ui/utils/responsive'
 import { formatNumber, getFractionDigitsOptions } from '@/ui/utils'
 import { shortenTokenAddress } from '@/utils'
 import { copyToClipboard } from '@/lib/utils'
-import dayjs from '@/lib/dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 import Box from '@/ui/Box'
 import { StyledIconButton, StyledInformationSquare16 } from '@/components/PagePool/PoolDetails/PoolStats/styles'
 import { Chip } from '@/ui/Typography'
