+++ apps/main/src/hooks/usePoolTotalStaked.tsx
@@ -2,7 +2,7 @@ import { Contract, Interface, JsonRpcProvider } from 'ethers'
 import { useCallback, useEffect } from 'react'
 import { isValidAddress } from '@/utils'
 import useStore from '@/store/useStore'
-import dayjs from '@/lib/dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 
 const usePoolTotalStaked = (poolDataCacheOrApi: PoolDataCacheOrApi) => {
   const { address, lpToken, gauge } = poolDataCacheOrApi?.pool ?? {}
