+++ apps/main/src/pages/_app.tsx
@@ -16,7 +16,7 @@ import { REFRESH_INTERVAL } from '@/constants'
 import GlobalStyle from '@/globalStyle'
 import usePageVisibleInterval from '@/hooks/usePageVisibleInterval'
 import Page from '@/layout/default'
-import { dynamicActivate, initTranslation, updateAppLocale } from '@/lib/i18n'
+import { dynamicActivate, initTranslation, updateAppLocale } from '@ui-kit/lib/i18n'
 import { messages as messagesEn } from '@/locales/en/messages.js'
 import curvejsApi from '@/lib/curvejs'
 import useStore from '@/store/useStore'
@@ -90,7 +90,10 @@ function CurveApp({ Component }: AppProps) {
     const { rLocale } = getLocaleFromUrl()
     const parsedLocale = rLocale?.value ?? 'en'
     initTranslation(i18n, parsedLocale)
-    dynamicActivate(parsedLocale)
+    ;(async () => {
+      let data = await import(`@/locales/${parsedLocale}/messages`)
+      dynamicActivate(parsedLocale, data)
+    })()
     updateAppLocale(parsedLocale, updateGlobalStoreByKey)
     ;(async () => {
       const networks = await fetchNetworks()
