+++ apps/main/src/store/createGlobalSlice.ts
@@ -1,6 +1,6 @@
 import type { GetState, SetState } from 'zustand'
 import type { State } from '@/store/useStore'
-import type { Locale } from '@/lib/i18n'
+import type { Locale } from '@ui-kit/lib/i18n'
 import { ConnectState, getPageWidthClassName } from '@/ui/utils'
 import isEqual from 'lodash/isEqual'
 import produce from 'immer'
