+++ apps/main/src/utils/storage.ts
@@ -1,7 +1,7 @@
 import type { Theme } from '@/store/createGlobalSlice'
 
 import merge from 'lodash/merge'
-import dayjs from '@/lib/dayjs'
+import dayjs from '@ui-kit/lib/dayjs'
 
 export const APP_STORAGE = {
   APP_CACHE: 'curve-app-cache',
