+++ packages/curve-ui-kit/src/lib/dayjs.ts
@@ -1,10 +1,10 @@
-import type { Locale } from '@/lib/i18n'
-
 import dayjs from 'dayjs'
 import isLeapYear from 'dayjs/plugin/isLeapYear'
 import utc from 'dayjs/plugin/utc'
 import customParseFormat from 'dayjs/plugin/customParseFormat'
 import localizedFormat from 'dayjs/plugin/localizedFormat'
+
+import type { Locale } from './i18n'
 
 dayjs.extend(isLeapYear)
 dayjs.extend(utc)
