+++ packages/curve-ui-kit/src/lib/i18n.ts
@@ -1,8 +1,9 @@
-import type { DefaultStateKeys } from '@/store/createAppSlice'
-
 import { I18n, i18n } from '@lingui/core'
 import { en, zh } from 'make-plural/plurals'
-import { setStorageValue } from '@/utils/storage'
+import numbro from 'numbro'
+import 'numbro/dist/languages.min.js'
+
+import { setDayjsLocale } from './dayjs'
 
 export type Locale = {
   name: string
@@ -10,14 +11,14 @@ export type Locale = {
   lang: string
 }
 
-export const DEFAULT_LOCALES: Locale[] = [{ name: 'English', value: 'en', lang: 'en' }]
+export const DEFAULT_LOCALES: Locale[] = [
+  { name: 'English', value: 'en', lang: 'en' },
+  { name: '简体中文', value: 'zh-Hans', lang: 'zh-Hans' },
+  { name: '繁體中文', value: 'zh-Hant', lang: 'zh-Hant' },
+]
 
 if (process.env.NODE_ENV === 'development') {
-  ;[
-    { name: '简体中文', value: 'zh-Hans' as const, lang: 'zh-Hans' },
-    { name: '繁體中文', value: 'zh-Hant' as const, lang: 'zh-Hant' },
-    { name: 'pseudo', value: 'pseudo' as const, lang: 'en' },
-  ].map((l) => DEFAULT_LOCALES.push(l))
+  DEFAULT_LOCALES.push({ name: 'pseudo', value: 'pseudo', lang: 'en' })
 }
 
 export function initTranslation(i18n: I18n, defaultLocale: string): void {
@@ -29,6 +30,10 @@ export function initTranslation(i18n: I18n, defaultLocale: string): void {
   })
   i18n.load(defaultLocale, {})
   i18n.activate(defaultLocale)
+}
+
+export function isLocaleInChinese(locale: Locale['value']) {
+  return locale === 'zh-Hant' || locale === 'zh-Hans'
 }
 
 export function findLocale(selectedLocale: string) {
@@ -49,14 +54,24 @@ export function parseLocale(locale?: string): { parsedLocale: Locale['value']; p
   }
 }
 
-export async function dynamicActivate(locale: string) {
-  const { parsedLocale } = parseLocale(locale)
-  let data = await import(`../locales/${parsedLocale}/messages`)
+export async function dynamicActivate(locale: string, data: any) {
   i18n.load(locale, data.messages)
   i18n.activate(locale)
 }
 
-export function updateAppLocale(locale: string, updateGlobalStoreByKey: <T>(key: DefaultStateKeys, value: T) => void) {
+export function updateAppLocale(locale: string, updateGlobalStoreByKey: <T>(key: 'locale', value: T) => void) {
   updateGlobalStoreByKey('locale', locale)
-  setStorageValue('APP_CACHE', { locale })
+  setDayjsLocale(locale as Locale['value'])
+
+  let numbroLang = ''
+  if (locale === 'zh-Hant') numbroLang = 'zh-TW'
+  if (locale === 'zh-Hans') numbroLang = 'zh-CN'
+
+  if (numbroLang) {
+    // @ts-ignore
+    import(`numbro/languages/${numbroLang}`).then((module) => {
+      numbro.registerLanguage(module.default)
+      numbro.setLanguage(numbroLang)
+    })
+  }
 }
